requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
pywebview

# Not needed anymore (using SQLite instead)
# psycopg2-binary==2.9.9
# python-dotenv==1.0.0
